% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointExtraction.R
\name{boxplot_points}
\alias{boxplot_points}
\title{boxplot_points}
\usage{
boxplot_points(axis_coords)
}
\arguments{
\item{axis_coords}{List output from cal_coords}
}
\value{
data.frame
}
\description{
Takes points user defined points from a boxplot, in a set order, and returns them.
}
\author{
Joel Pick
}
