% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_metaDigitise.R
\name{bulk_edit}
\alias{bulk_edit}
\title{bulk_edit}
\usage{
bulk_edit(dir, summary = TRUE)
}
\arguments{
\item{dir}{parent directory}

\item{summary}{logical; whether summary is returned}
}
\description{
Function for bulk editing previous data extraction through `metaDigitise`
}
\author{
Joel Pick
}
