% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_coords.R
\name{cal_coords}
\alias{cal_coords}
\title{cal_coords}
\usage{
cal_coords(plot_type, cex, ...)
}
\arguments{
\item{plot_type}{plot type}

\item{cex}{size of points}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Prompts user to enter axis coordinates, and their values. Modified from the digitize package
}
