% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_coords.R
\name{calibrate}
\alias{calibrate}
\title{calibrate}
\usage{
calibrate(raw_data, calpoints, point_vals)
}
\description{
Converts x and y coordinates from original plot coords to actual coords using previous identified coordinates. Modified from digitise package
}
