% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_func.R
\name{cat_matrix}
\alias{cat_matrix}
\title{cat_matrix}
\usage{
cat_matrix(x, cols)
}
\arguments{
\item{x}{vector}

\item{cols}{number of columns}
}
\description{
prints a vector as a number list of items with a certain number of columns
}
