% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformStats.R
\name{CI95_to_sd}
\alias{CI95_to_sd}
\title{CI95_to_sd}
\usage{
CI95_to_sd(CI, n)
}
\arguments{
\item{CI}{Interval difference from the mean}

\item{n}{Sample Size}
}
\value{
Returns vector of standard deviations
}
\description{
Transforms symmetrical confidence interval to standard deviation
}
\examples{
CI95_to_sd(CI = 2, n = 10)
}
\author{
Joel Pick
}
