% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{convert_histogram_data}
\alias{convert_histogram_data}
\title{convert_histogram_data}
\usage{
convert_histogram_data(cal_data)
}
\arguments{
\item{cal_data}{The calibration data}
}
\description{
Conversion of extracted data from histogram
}
