% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDigitise.R
\name{dir_details}
\alias{dir_details}
\title{dir_details}
\usage{
dir_details(dir)
}
\arguments{
\item{dir}{the path name to the directory / folder where the files are located}
}
\description{
Function will gather important directory details about calibration files and figures needed for processing
}
\author{
Daniel Noble - daniel.wa.noble@gmail.com
}
