% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_metaDigitise.R
\name{enter_N}
\alias{enter_N}
\title{enter_N}
\usage{
enter_N(raw_data, ...)
}
\arguments{
\item{raw_data}{raw_data}

\item{...}{Pass additional arguments}
}
\description{
Enter sample sizes for a group
}
\author{
Joel Pick
}
