% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_metaDigitise.R
\name{extract_digitised}
\alias{extract_digitised}
\title{extract_digitised}
\usage{
extract_digitised(list)
}
\arguments{
\item{list}{A list of objects returned from metaDigitise}
}
\value{
The function will return a data frame with the data across all the digitised files
}
\description{
Function for extracting the data from a metaDigitise list
}
