% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_points.R
\name{extract_points}
\alias{extract_points}
\title{extract_points}
\usage{
extract_points(file, plot_type = c("mean_se", "boxplot", "scatterplot"),
  summary_stats = FALSE)
}
\arguments{
\item{file}{Image file}

\item{plot_type}{One of "mean_se", "boxplot", or "scatterplot". "mean_se" assumes that there are means and standard error bars, and requires the user to click the upper error bar followed by the mean. "boxplot" assumes that the user will input 5 points, in the order max, upper (75th) quartile, median, lower (25th) quartile, and minimum. "Scatterplot assumes that all points will be clicked"}

\item{summary_stats}{Whether further summary statistics are derived from "mean_se" and "boxplot". Require use to input sample sizes. Currently defunct}
}
\value{
Dataframe
}
\description{
Extracts points from a figure and generate summary statistics
}
\author{
Joel Pick
}
