% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nearest_point.R
\name{find_nearest_point}
\alias{find_nearest_point}
\title{find_nearest_point}
\usage{
find_nearest_point(raw_data, image_width, image_height)
}
\arguments{
\item{raw_data}{raw_data}

\item{image_width}{image_width}

\item{image_height}{image_height}
}
\description{
find nearest point
}
