% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDigitise.R
\name{get_notDone_file_details}
\alias{get_notDone_file_details}
\title{get_notDone_file_details}
\usage{
get_notDone_file_details(dir)
}
\arguments{
\item{dir}{Path name to the directory / folder where the figure files are located.}
}
\value{
Returns a list containing details on the images names and their paths, the calibration file names (or files already completed) as well as the paths to these files.
}
\description{
Function will get file information from the directory and the calibration files. It will also exclude files that have already been processed, as is judged by the match between file names in the calibration folder and the imported details object
}
\author{
Daniel Noble - daniel.wa.noble@gmail.com
}
