% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_metaDigitise.R
\name{getExtracted}
\alias{getExtracted}
\title{getExtracted}
\usage{
getExtracted(dir, summary = TRUE)
}
\arguments{
\item{dir}{The directory where figures have already been digitised. There}

\item{summary}{Logical indicating whether summarised (default) or calibrated data should be returned.}
}
\value{
Returns a data frame (summary = TRUE) or a list with slots for each plot type (summary = FALSE)
}
\description{
Extracts data from a directory that has been previous digitised using metaDigitise
}
