% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_coords.R
\name{getVals}
\alias{getVals}
\title{getVals}
\usage{
getVals(calpoints, image_details, ...)
}
\arguments{
\item{calpoints}{Calibration points}

\item{image_details}{image width and height}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Gets values needed to calibrate axis coordinated. Modified from the digitize package
}
