% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformStats.R
\name{grandMean}
\alias{grandMean}
\title{grandMean}
\usage{
grandMean(mean, n)
}
\arguments{
\item{mean}{Mean}

\item{n}{Sample size}
}
\value{
Returns vector of pooled mean
}
\description{
Pooled mean of a set of group means
}
\examples{
grandMean(mean = 10, n = 30)
}
\author{
Joel Pick
}
