% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformStats.R
\name{grandSD}
\alias{grandSD}
\title{grandSD}
\usage{
grandSD(mean, sd, n, equal = FALSE)
}
\arguments{
\item{mean}{Mean}

\item{sd}{standard deviation}

\item{n}{Sample size}

\item{equal}{Logical: Whether to calculate pooled SD assuming groups have the same means (TRUE) or different means (FALSE)}
}
\value{
Returns vector of pooled mean
}
\description{
Pooled standard deviation of a set of groups
}
\examples{
grandSD(mean = 10, sd = 3, n = 40)
}
\author{
Joel Pick
}
