% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_rotate.R
\name{graph_rotate}
\alias{graph_rotate}
\title{graph_rotate}
\usage{
graph_rotate(image)
}
\arguments{
\item{image}{Image object from magick::image_read}
}
\value{
Returns new image object and plots rotated image
}
\description{
Rotates imported figures in order to align them properly. Asks the user after each rotation if further rotation is required
}
\author{
Joel Pick
}
