% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ME_Box_extraction.R
\name{groups_extract}
\alias{groups_extract}
\title{groups_extract}
\usage{
groups_extract(plot_type, nGroups)
}
\description{
Extraction of data from boxplots of mean_error plots, from multiple groups
}
