% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_extract.R
\name{histogram_extract}
\alias{histogram_extract}
\title{histogram_extract}
\usage{
histogram_extract(image_file, calpoints, point_vals, ...)
}
\arguments{
\item{image_file}{image file name}

\item{calpoints}{points used for calibration}

\item{point_vals}{values for calibration}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Extraction of data from histograms
}
