% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_metaDigitise.R
\name{import_menu}
\alias{import_menu}
\title{import_menu}
\usage{
import_menu(dir, summary)
}
\arguments{
\item{dir}{The directory where figures have already been digitised}

\item{summary}{Logical indicating whether the imported data should be returned in summarised or processed form.}
}
\value{
Returns a list (summary = FALSE) or data frame (summary = TRUE)
}
\description{
Imports metaDigitise() calibration files from a directory that is partially or fully digitised already
}
