% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_metaDigitise.R
\name{import_metaDigitise}
\alias{import_metaDigitise}
\title{import_metaDigitise}
\usage{
import_metaDigitise(dir, summary = TRUE)
}
\arguments{
\item{dir}{The directory where figures have already been digitised}

\item{summary}{Logical indicating whether the imported data should be returned in summarised form ('TRUE') or not ('FALSE')}
}
\value{
Returns a list if types = 'diff' or a dataframe of raw data or summarise data if types = 'same' or summary = TRUE
}
\description{
Imports metaDigitise calibration files from a directory that is partially or fully digitised already
}
