% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_func.R
\name{is.even}
\alias{is.even}
\title{is.even}
\usage{
is.even(x)
}
\arguments{
\item{x}{integer value}
}
\value{
Logical (TRUE or FALSE) indicating whether value is an even number or not
}
\description{
Checks whether a integer is even
}
