% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_func.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{is.wholenumber}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{object to be  tested}

\item{tol}{tolerance}
}
\value{
Logical value (TRUE or FALSE)
}
\description{
Checks whether value is a whole number
}
