% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_func.R
\name{isNumeric}
\alias{isNumeric}
\title{isNumeric}
\usage{
isNumeric(x)
}
\arguments{
\item{x}{character to be tested}
}
\value{
Logical (TRUE or FALSE) indicating whether value is numeric or not
}
\description{
Checks whether a character is a number
}
