% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_func.R
\name{knownN}
\alias{knownN}
\title{knownN}
\usage{
knownN(plot_type, processed_data, knownN = NULL, ...)
}
\arguments{
\item{plot_type}{plot type}

\item{processed_data}{raw_data}

\item{knownN}{previously entered N}

\item{...}{arguments from other calls}
}
\description{
prints a vector as a number list of items with a certain number of columns
}
