% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_metaDigitise.R
\name{load_metaDigitise}
\alias{load_metaDigitise}
\title{load_metaDigitise}
\usage{
load_metaDigitise(doneCalFiles, names)
}
\arguments{
\item{doneCalFiles}{The calibration files that have already been finished taken from directory details}

\item{names}{The names of the done calibration files}
}
\value{
Returns a list of metaDigitised objects that have already been completed
}
\description{
Loads metaDigitise calibration files from a directory that is partially or fully digitised already
}
