% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_func.R
\name{locator_mD}
\alias{locator_mD}
\title{locator_mD}
\usage{
locator_mD(nPoints = 1, line = TRUE, cex = 1, col = "red", ...)
}
\arguments{
\item{nPoints}{number of points in a sequence}

\item{line}{logical; plot lines between points}

\item{cex}{size of points}

\item{col}{colour of points}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Plots clicked points, and returns their x.y coordinates as a data.frame
}
\description{
Wrapper function for locator, with more control over point size etc
}
