% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_coords.R
\name{logAxes}
\alias{logAxes}
\title{getVals}
\usage{
logAxes(...)
}
\arguments{
\item{...}{further arguments passed to or from other methods.}
}
\description{
Ask user for information about whether axes are on log scale
}
