% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointExtraction.R
\name{mean_se_points}
\alias{mean_se_points}
\title{mean_se_points}
\usage{
mean_se_points(axis_coords)
}
\arguments{
\item{axis_coords}{List output from cal_coords}
}
\value{
data.frame
}
\description{
Takes points user defined points for mean and se, in a set order, and returns them.
}
\author{
Joel Pick
}
