% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDigitise.R
\name{metaDigitise}
\alias{metaDigitise}
\title{metaDigitise}
\usage{
metaDigitise(image_file, plot_type = NULL, summary_stats = FALSE)
}
\arguments{
\item{image_file}{Image file}

\item{plot_type}{Type of plot from "mean_error","boxplot","scatterplot" or"histogram". Function will prompt if not entered by user.}

\item{summary_stats}{Whether further summary statistics are derived from "mean_se" and "boxplot". Require user to input sample sizes. Currently defunct}
}
\value{
List of
}
\description{
Extracts points from a figure and generate summary statistics
}
\author{
Joel Pick
}
