% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_metaDigitise.R
\name{order_lists}
\alias{order_lists}
\title{order_lists}
\usage{
order_lists(list, plot_types)
}
\arguments{
\item{list}{The calibration files that have already been finished}

\item{plot_types}{The list of plot types}
}
\value{
Returns a list ordered by the plot type
}
\description{
Will re-order the processed data such that similar types of data are organised into a single list defined by their
}
