% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDigitise.R
\name{plot.metaDigitise}
\alias{plot.metaDigitise}
\title{plot.metaDigitise}
\usage{
\method{plot}{metaDigitise}(x, ...)
}
\arguments{
\item{x}{an R object of class ‘metaDigitise’}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Re-plots figure and extraction data
}
\author{
Joel Pick
}
