% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_extraction.R
\name{point_extraction}
\alias{point_extraction}
\title{point_extraction}
\usage{
point_extraction(object, edit = FALSE)
}
\arguments{
\item{object}{Object}

\item{edit}{Logical (TRUE or FALSE) indicating whether a point would like to be edited}
}
\description{
Extracts or edits point of a digitisation
}
