% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDigitise.R
\name{print.metaDigitise}
\alias{print.metaDigitise}
\title{metaDigitise}
\usage{
\method{print}{metaDigitise}(x, ...)
}
\arguments{
\item{x}{an R object of class ‘metaDigitise’}
}
\description{
Extracts points from a figure and generate summary statistics
}
\author{
Joel Pick
}
