% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_coords.R
\name{print_cal_instructions}
\alias{print_cal_instructions}
\title{print_cal_instructions}
\usage{
print_cal_instructions(plot_type, ...)
}
\arguments{
\item{plot_type}{plot type}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Prints instructions for calibration. Modified from the digitize package
}
