% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{process_data}
\alias{process_data}
\title{process_data}
\usage{
process_data(object)
}
\arguments{
\item{object}{object from metaDigitise}
}
\description{
Processes points clicked into a meaningful dataframe
}
