% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDigitise.R
\name{process_new_files}
\alias{process_new_files}
\title{process_new_files}
\usage{
process_new_files(dir, summary = TRUE)
}
\arguments{
\item{dir}{the path name to the directory / folder where the files are located}

\item{summary}{summary = TRUE or FALSE is most relevant as it will print a simple summary statistics that are the same across all files}
}
\description{
Batch processes image files within a set directory, consolidates the data and exports the data for each image and type
}
\author{
Joel Pick - joel.l.pick@gmail.com

Daniel Noble - daniel.wa.noble@gmail.com
}
