% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformStats.R
\name{range_to_sd}
\alias{range_to_sd}
\title{range_to_sd}
\usage{
range_to_sd(min, max, n)
}
\arguments{
\item{min}{Minimum value}

\item{max}{Maximum value}

\item{n}{Sample size}
}
\value{
Returns vector of standard deviation
}
\description{
Converts a range to a standard deviation
}
\examples{
range_to_sd(min = 3, max = 8, n = 40)
}
\author{
Joel Pick
}
