% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDigitise.R
\name{redraw.metaDigitise}
\alias{redraw.metaDigitise}
\title{redraw.metaDigitise}
\usage{
redraw.metaDigitise(object)
}
\arguments{
\item{object}{an R object of class ‘metaDigitise’}
}
\description{
Redraws figure and extraction data
}
\author{
Joel Pick
}
