% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redraw.R
\name{redraw}
\alias{redraw}
\title{redraw}
\usage{
redraw(image, plot_type, raw_data, calpoints, point_vals)
}
\arguments{
\item{image}{image}

\item{plot_type}{The type of plot}

\item{raw_data}{previously extracted data points}

\item{calpoints}{points used for calibration}

\item{point_vals}{values for calibration}
}
\description{
Extraction of data from boxplots of mean_error plots, from multiple groups
}
