% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redraw.R
\name{redraw_calibration}
\alias{redraw_calibration}
\title{redraw_calibration}
\usage{
redraw_calibration(plot_type, calpoints, point_vals, image_details)
}
\arguments{
\item{plot_type}{plot_type}

\item{calpoints}{The calibration points}

\item{point_vals}{The point values}

\item{image_details}{image_details}
}
\description{
plots calibration data on graph
}
