% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redraw.R
\name{redraw_rotation}
\alias{redraw_rotation}
\title{rotate_graph}
\usage{
redraw_rotation(image, flip, rotate)
}
\arguments{
\item{image}{Image object from magick::image_read}

\item{flip}{whether to flip figure}

\item{rotate}{how much to rotate figure}
}
\description{
Rotates/flips imported figures
}
