% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_rotation.R
\name{rotate_graph}
\alias{rotate_graph}
\title{rotate_graph}
\usage{
rotate_graph(image, flip, rotate)
}
\arguments{
\item{image}{Image object from magick::image_read}

\item{flip}{whether to flip figure}

\item{rotate}{how much to rotate figure}
}
\description{
Rotates/flips imported figures
}
