% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformStats.R
\name{rqm_to_mean}
\alias{rqm_to_mean}
\title{rqm_to_mean}
\usage{
rqm_to_mean(min, LQ, median, UQ, max)
}
\arguments{
\item{min}{Minimum value}

\item{LQ}{Lower 75% quartile}

\item{median}{Median}

\item{UQ}{Upper 75% quartile}

\item{max}{Maximum value}
}
\value{
Returns vector of mean
}
\description{
Calculate the mean from the box plots
}
\author{
Joel Pick
}
