% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformStats.R
\name{se_to_sd}
\alias{se_to_sd}
\title{se_to_sd}
\usage{
se_to_sd(se, n, na.rm = TRUE)
}
\arguments{
\item{se}{Standard Error of the mean}

\item{n}{Sample Size}

\item{na.rm}{Remove NAs}
}
\value{
Returns vector of standard errors
}
\description{
Transforms standerd error to standard deviation
}
\author{
Joel Pick
}
