% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_metaDigitise.R
\name{setup_calibration_dir}
\alias{setup_calibration_dir}
\title{setup_calibration_dir}
\usage{
setup_calibration_dir(dir)
}
\arguments{
\item{dir}{the path name to the directory / folder where the files are located}
}
\description{
Function will check whether the calibration directory has been setup and if not, create one.
}
