% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ME_Box_extraction.R
\name{single_MB_extract}
\alias{single_MB_extract}
\title{single_MB_extract}
\usage{
single_MB_extract(plot_type, cex)
}
\arguments{
\item{plot_type}{Type of plot}

\item{cex}{point size}
}
\description{
Takes points user defined points from a single group mean_error plot or boxplot, in a set order, and returns them.
}
