% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_extract.R
\name{single_scatter_extract}
\alias{single_scatter_extract}
\title{single_scatter_extract}
\usage{
single_scatter_extract(image, calpoints, point_vals)
}
\arguments{
\item{image}{image}

\item{calpoints}{points used for calibration}

\item{point_vals}{values for calibration}
}
\description{
Extraction of data from scatterplots
}
