% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_metaDigitise.R
\name{specify_type}
\alias{specify_type}
\title{specify_type}
\usage{
specify_type()
}
\description{
Function that allows user to interface with function to specific each type of plot prior to digitising
}
