% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDigitise.R
\name{summary.metaDigitise}
\alias{summary.metaDigitise}
\title{metaDigitise}
\usage{
\method{summary}{metaDigitise}(object, ...)
}
\arguments{
\item{object}{an R object of class ‘metaDigitise’}
}
\value{
Data.frame
}
\description{
Extracts points from a figure and generate summary statistics
}
\author{
Joel Pick
}
