% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_func.R
\name{user_base}
\alias{user_base}
\title{user_base}
\usage{
user_base(...)
}
\arguments{
\item{...}{arguments passed to other functions}
}
\description{
asks user for base of logarithm, accept numeric or "e"
}
