% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_coords.R
\name{user_calibrate}
\alias{user_calibrate}
\title{user_calibrate}
\usage{
user_calibrate(object)
}
\arguments{
\item{object}{metaDigitise object}
}
\description{
Gets values needed to calibrate axis coordinated. Modified from the digitize package
}
