% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_func.R
\name{user_options}
\alias{user_options}
\title{user_options}
\usage{
user_options(question, allowed_answers)
}
\arguments{
\item{question}{question}

\item{allowed_answers}{allowed answers}
}
\description{
asks user for option from specified list
}
