% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_func.R
\name{user_unique}
\alias{user_unique}
\title{user_unique}
\usage{
user_unique(question, previous_answers)
}
\arguments{
\item{question}{question}

\item{previous_answers}{allowed answers}
}
\description{
asks user for option from specified list
}
