% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_metaDigitise.R
\name{bulk_metaDigitise}
\alias{bulk_metaDigitise}
\title{bulk_digitise}
\usage{
bulk_metaDigitise(dir, types = c("diff", "same"), ...)
}
\arguments{
\item{dir}{the path name to the directory / folder where the files are located}

\item{types}{Argument specifying whether the types of images are the same (i.e., all scatter plots - "same") or a mixture of different plot types (i.e., scatter plots, means and se - "diff").}

\item{...}{Other arguments called from metaDigitise. summary = TRUE or FALSE is most relevant as it will print a simple summary statistics that are the same across all files.}
}
\value{
If type = "same" the function returns a dataframe with the relevant data for each figure being digitised. If type = "diff" it returns a list of the relevant data.
}
\description{
Batch processes png files within a set directory, consolidates the data and exports the data for each image and type
}
