% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{calibrate}
\alias{calibrate}
\title{calibrate}
\usage{
calibrate(raw_data, calpoints, point_vals, log_axes, ...)
}
\arguments{
\item{raw_data}{The raw data}

\item{calpoints}{The calibration points}

\item{point_vals}{The point values}

\item{log_axes}{whether x or y is logged}

\item{...}{further arguments passed to or from other methods}
}
\description{
Converts x and y coordinates from original plot coords to actual coords using previous identified coordinates. Modified from digitise package
}
