% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ME_Box_extraction.R
\name{convert_group_data}
\alias{convert_group_data}
\title{convert_group_data}
\usage{
convert_group_data(cal_data, plot_type, nGroups)
}
\arguments{
\item{cal_data}{Calibrated data}

\item{plot_type}{The type of plot}

\item{nGroups}{number of groups}
}
\description{
Converts, pre-calibrated points clicked into a meaningful dataframe
}
