% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_extract.R
\name{edit_group}
\alias{edit_group}
\title{edit_group}
\usage{
edit_group(raw_data, group_id, calpoints, cex, ...)
}
\arguments{
\item{raw_data}{data}

\item{group_id}{group_id}

\item{calpoints}{The calibration points}

\item{cex}{point size}

\item{...}{other functions to pass to internal_redraw}
}
\description{
Edit group points in scatterplots
}
