% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_metaDigitise.R
\name{edit_metaDigitise}
\alias{edit_metaDigitise}
\title{edit_metaDigitise}
\usage{
edit_metaDigitise(object)
}
\arguments{
\item{object}{an R object of class ‘metaDigitise’}
}
\value{
Data.frame
}
\description{
Function for editing previous data extraction through `metaDigitise`
}
\author{
Joel Pick
}
