% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDigitise.R
\name{extract_digitised}
\alias{extract_digitised}
\title{extract_digitised}
\usage{
extract_digitised(list, types = c("same", "diff"), summary = TRUE)
}
\arguments{
\item{list}{A list of objects returned from metaDigitise}

\item{types}{Whether the plots are the same or different types}

\item{summary}{A logical 'TRUE' or 'FALSE' indicating whether metaDigitise should print summary statitics from each figure and group.}
}
\value{
The function will return a data frame with the data across all the digitised files
}
\description{
Function for extracting the data from a metaDigitise list
}
