% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_metaDigitise.R
\name{get_notDone_file_details}
\alias{get_notDone_file_details}
\title{get_notDone_file_details}
\usage{
get_notDone_file_details(dir)
}
\arguments{
\item{dir}{the path name to the directory / folder where the files are located}
}
\description{
Function will get a series of file information from the directory and the calibration files. It will also exclude files that have already been processed, as is judged by the match between file names in the calibration folder and the
}
