% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_coords.R
\name{getVals}
\alias{getVals}
\title{getVals}
\usage{
getVals(calpoints, image_width, image_height)
}
\arguments{
\item{calpoints}{Calibration points}

\item{image_width}{image width}

\item{image_height}{image height}
}
\description{
Gets values needed to calibrate axis coordinated. Modified from the digitize package
}
