% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_extract.R
\name{group_scatter_extract}
\alias{group_scatter_extract}
\title{group_scatter_extract}
\usage{
group_scatter_extract(edit = FALSE.nGroups, ...)
}
\arguments{
\item{...}{arguments passed to internal_redraw}

\item{nGroups}{The number of groups}
}
\description{
Extraction of data from scatterplots
}
