% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_extract.R
\name{group_scatter_extract}
\alias{group_scatter_extract}
\title{group_scatter_extract}
\usage{
group_scatter_extract(nGroups, image, image_file, calpoints, point_vals)
}
\arguments{
\item{nGroups}{The number of groups}

\item{image}{image}

\item{image_file}{image file name}

\item{calpoints}{points used for calibration}

\item{point_vals}{values for calibration}
}
\description{
Don't know yet Joel.
}
