% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_extract.R
\name{group_scatter_extract}
\alias{group_scatter_extract}
\title{group_scatter_extract}
\usage{
group_scatter_extract(edit = FALSE, raw_data = data.frame(), cex, ...)
}
\arguments{
\item{edit}{logical; whether in edit mode}

\item{raw_data}{raw data}

\item{cex}{point size}

\item{...}{arguments passed to internal_redraw}
}
\description{
Extraction of data from scatterplots
}
