% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ME_Box_extraction.R
\name{groups_extract}
\alias{groups_extract}
\title{groups_extract}
\usage{
groups_extract(plot_type, nGroups, image, calpoints, point_vals)
}
\arguments{
\item{plot_type}{The type of plot}

\item{nGroups}{number of groups}

\item{image}{image}

\item{calpoints}{points used for calibration}

\item{point_vals}{values for calibration}
}
\description{
Extraction of data from boxplots of mean_error plots, from multiple groups
}
