% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_extract.R
\name{histogram_extract}
\alias{histogram_extract}
\title{histogram_extract}
\usage{
histogram_extract(image, image_file, calpoints, point_vals)
}
\arguments{
\item{image}{image}

\item{image_file}{image file name}

\item{calpoints}{points used for calibration}

\item{point_vals}{values for calibration}
}
\description{
Extraction of data from histograms
}
