% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_extract.R
\name{histogram_extract}
\alias{histogram_extract}
\title{histogram_extract}
\usage{
histogram_extract(...)
}
\arguments{
\item{...}{arguments to pass to internal_redraw}
}
\description{
Extraction of data from histograms
}
