% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_extract.R
\name{histogram_extract}
\alias{histogram_extract}
\title{histogram_extract}
\usage{
histogram_extract(edit = FALSE, raw_data = data.frame(), calpoints, cex,
  ...)
}
\arguments{
\item{edit}{logical; whether in edit mode}

\item{raw_data}{raw data}

\item{calpoints}{The calibration points}

\item{cex}{point size}

\item{...}{arguments to pass to internal_redraw}
}
\description{
Extraction of data from histograms
}
