% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_metaDigitise.R
\name{import_metaDigitise}
\alias{import_metaDigitise}
\title{import_metaDigitise}
\usage{
import_metaDigitise(dir, summary)
}
\arguments{
\item{dir}{The directory where figures have already been digitised}

\item{summary}{Logical indicating whether the imported data should be returned in summarised form ('TRUE') or not ('FALSE')}
}
\value{
Returns a list (summary = FALSE) or data frame (summary = TRUE)
}
\description{
Imports metaDigitise() calibration files from a directory that is partially or fully digitised already
}
\author{
Daniel Noble - daniel.wa.noble@gmail.com
}
