% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redraw.R
\name{internal_redraw}
\alias{internal_redraw}
\title{internal_redraw}
\usage{
internal_redraw(image, plot_type, calpoints, point_vals, raw_data)
}
\arguments{
\item{image}{image}

\item{plot_type}{The type of plot}

\item{calpoints}{points used for calibration}

\item{point_vals}{values for calibration}

\item{raw_data}{previously extracted data points}
}
\description{
Redraws figure and extraction data
}
\author{
Joel Pick
}
