% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_metaDigitise.R
\name{load_metaDigitise}
\alias{load_metaDigitise}
\title{load_metaDigitise}
\usage{
load_metaDigitise(doneCalFiles, names)
}
\arguments{
\item{doneCalFiles}{The metaDigitise objects that have already been completed in the directory}

\item{names}{The names of the finished metaDigitise objects}
}
\value{
Returns a list of metaDigitised objects that have already been completed
}
\description{
Loads metaDigitise calibration / data files from a directory containing a set of figures that are partially or fully digitised already.
}
\author{
Daniel Noble - daniel.wa.noble@gmail.com
}
