% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ME_Box_extraction.R
\name{MB_extract}
\alias{MB_extract}
\title{groups_extract}
\usage{
MB_extract(edit = FALSE, plot_type, entered_N, raw_data = data.frame(), ...)
}
\arguments{
\item{plot_type}{The type of plot}

\item{entered_N}{ask for sample sizes?}

\item{...}{further arguments to internal_redraw}

\item{nGroups}{number of groups}
}
\description{
Extraction of data from boxplots of mean_error plots, from multiple groups
}
