% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ME_Box_extraction.R
\name{MB_extract}
\alias{MB_extract}
\title{groups_extract}
\usage{
MB_extract(plot_type, nGroups, image_file, image_details, calpoints, point_vals,
  entered_N, ...)
}
\arguments{
\item{plot_type}{The type of plot}

\item{nGroups}{number of groups}

\item{image_file}{image file name}

\item{image_details}{image details}

\item{calpoints}{points used for calibration}

\item{point_vals}{values for calibration}

\item{entered_N}{ask for sample sizes?}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Extraction of data from boxplots of mean_error plots, from multiple groups
}
