% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ME_Box_extraction.R
\name{MB_extract}
\alias{MB_extract}
\title{MB_extract}
\usage{
MB_extract(edit = FALSE, plot_type, entered_N,
  raw_data = data.frame(stringsAsFactors = TRUE), cex, ...)
}
\arguments{
\item{edit}{logical; whether in edit mode}

\item{plot_type}{The type of plot}

\item{entered_N}{ask for sample sizes?}

\item{raw_data}{raw data}

\item{cex}{point size}

\item{...}{further arguments to MB_extract}
}
\description{
Extraction of data from boxplots of mean_error plots, from multiple groups
}
