% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDigitise.R
\name{metaDigitise}
\alias{metaDigitise}
\title{metaDigitise}
\usage{
metaDigitise(image_file, plot_type = NULL)
}
\arguments{
\item{image_file}{Image file}

\item{plot_type}{Type of plot from "mean_error","boxplot","scatterplot" or"histogram". Function will prompt if not entered by user.}
}
\value{
List of
}
\description{
Extracts points from a figure and generate summary statistics
}
\author{
Joel Pick
}
