% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_points.R
\name{metaDigitise}
\alias{metaDigitise}
\title{metaDigitise}
\usage{
metaDigitise(file, summary_stats = FALSE)
}
\arguments{
\item{file}{Image file}

\item{summary_stats}{Whether further summary statistics are derived from "mean_se" and "boxplot". Require use to input sample sizes. Currently defunct}
}
\value{
Dataframe
}
\description{
Extracts points from a figure and generate summary statistics
}
\author{
Joel Pick
}
