% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDigitise.R
\name{metaDigitise}
\alias{metaDigitise}
\title{metaDigitise}
\usage{
metaDigitise(dir, summary = TRUE)
}
\arguments{
\item{dir}{the path name to the directory / folder where the files are located}

\item{summary}{whether the digitised data should be returned as a summary (TRUE) or as a concatenated list of similar types.}
}
\value{
A data frame or list containing the raw digitised data or the processed, summary statistics from the digitised data
}
\description{
Single or batch processes figures of .png, .jpg, .tiff, .pdf extensions within a set directory, consolidates the data and exports the data for each image and type
}
\details{
metaDigitise can be used on a directory with a whole host of different figure (mean and error, scatter plots, box plots and histograms) and file types (.jpeg, .png, .tiff, .pdf). It will automatically cycle through all files within a directory in order, prompting the user for specific information as they go. It will also write calibration files (also containing processed data), into a special caldat/ folder within the directory. Importantly, as new files are added to a directory that has already been "completed", metaDigitise will recognize these unfinished files and only cycle through the digitisation of these new files.
}
\examples{
# data <- metaDigitise(dir = "./example_figs/", summary = TRUE)
# summary(data)
}
