% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDigitise_functions.R
\name{print.metaDigitise}
\alias{print.metaDigitise}
\title{print.metaDigitise}
\usage{
\method{print}{metaDigitise}(x, ...)
}
\arguments{
\item{x}{an R object of class ‘metaDigitise’}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print method for class ‘metaDigitise’
}
\author{
Joel Pick
}
