% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redraw.R
\name{redraw_calibration}
\alias{redraw_calibration}
\title{redraw_calibration}
\usage{
redraw_calibration(plot_type, variable, calpoints, point_vals,
  image_details, cex)
}
\arguments{
\item{plot_type}{plot_type}

\item{variable}{variable}

\item{calpoints}{The calibration points}

\item{point_vals}{The point values}

\item{image_details}{image_details}

\item{cex}{relative size of points and text}
}
\description{
plots calibration data on graph
}
