% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redraw.R
\name{redraw_points}
\alias{redraw_points}
\title{redraw_points}
\usage{
redraw_points(plot_type, raw_data, image_details)
}
\arguments{
\item{plot_type}{plot_type}

\item{raw_data}{The raw data}

\item{image_details}{image_details}
}
\description{
plots clicked data on graph
}
