% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformStats.R
\name{se_to_sd}
\alias{se_to_sd}
\title{se_to_sd}
\usage{
se_to_sd(se, n)
}
\arguments{
\item{se}{Standard Error of the mean}

\item{n}{Sample Size}
}
\value{
Returns vector of standard errors
}
\description{
Transforms standard error to standard deviation
}
\examples{
se_to_sd(se = 5, n = 10)
}
\author{
Joel Pick
}
