% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDigitise.R
\name{setup_calibration_dir}
\alias{setup_calibration_dir}
\title{setup_calibration_dir}
\usage{
setup_calibration_dir(dir)
}
\arguments{
\item{dir}{Path name to the directory / folder where the files are located.}
}
\value{
Returns a caldat/ folder within the directory where all metaDigitise objects are stored.
}
\description{
Function will check whether the calibration directory has been setup and if not, create one.
}
\examples{
\donttest{
# temporary directory
tmp_dir <- tempdir()

#Create the calibration folder in the directory specified that is used to store files.
setup_calibration_dir(paste0(tmp_dir, "/"))

}
}
\author{
Daniel Noble - daniel.wa.noble@gmail.com
}
