% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDigitise.R
\name{specify_type}
\alias{specify_type}
\title{specify_type}
\usage{
specify_type()
}
\value{
The function will return the type of plot specified by the user and feed this argument back into metDigitise
}
\description{
Function that allows user to interface with function to specific each type of plot prior to digitising
}
\author{
Daniel Noble - daniel.wa.noble@gmail.com

Joel Pick - joel.l.pick@gmail.com
}
