% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDigitise_functions.R
\name{summary.metaDigitise}
\alias{summary.metaDigitise}
\title{summary.metaDigitise}
\usage{
\method{summary}{metaDigitise}(object, ...)
}
\arguments{
\item{object}{an R object of class ‘metaDigitise’}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Data.frame
}
\description{
Summary method for class ‘metaDigitise’
}
\author{
Joel Pick
}
