% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_rotation.R
\name{user_rotate_graph}
\alias{user_rotate_graph}
\title{user_rotate_graph}
\usage{
user_rotate_graph(image_file)
}
\arguments{
\item{image_file}{Image filename}
}
\description{
Rotates/flips imported figures according to user input, in order to align them properly. Asks the user after each change if further alteration is required
}
